# swagger-ui/api-docs.yaml
openapi: 3.0.3
info:
  title: Todo App REST API
  description: |
    A comprehensive REST API for a Todo application built with Express.js and JSON Server.
    
    ## Features
    - User registration and authentication
    - Todo CRUD operations
    - Category management
    - User-Todo relationships
    - JWT-based authentication
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api-proxy
    description: Proxied API Server (Recommended for Codespaces)
  - url: http://localhost:5000
    description: Local Development Express Server
  - url: http://localhost:3001
    description: Local Development JSON Server
  - url: http://localhost:3001
    description: JSON Server (Database)
...
paths:
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get categories
      description: Get all categories for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

    post:
      tags:
        - Categories
      summary: Create category
      description: Create a new category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category created successfully"
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      description: Get a specific category by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Category ID
          schema:
            type: string
      responses:
        '200':
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'

    put:
      tags:
        - Categories
      summary: Update category
      description: Update a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Category ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category updated successfully"
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'

    delete:
      tags:
        - Categories
      summary: Delete category
      description: Delete a category (cannot delete default categories or categories in use)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Category ID
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # JSON Server Direct Endpoints (for testing)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          example: "1"
        userId:
          type: string
          example: "1"
        name:
          type: string
          example: "Work"
        description:
          type: string
          example: "Work-related tasks"
        color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$'
          example: "#3498db"
        icon:
          type: string
          example: "💼"
        isDefault:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Work"
        description:
          type: string
          maxLength: 200
          example: "Work-related tasks"
        color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$'
          example: "#3498db"
        icon:
          type: string
          maxLength: 10
          example: "💼"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
...